Move overrided service to JobServlet
