--- walk-in-interview/src/main/java/com/google/job/servlets/MyServlet.java
@@ -1,25 +1,0 @@
-package com.google.job.servlets;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-/** Class to handle PATCH request. */
-public abstract class MyServlet extends HttpServlet {
-    private static final String PATCH = "PATCH";
-
-    @Override
-    public void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        if (request.getMethod().equalsIgnoreCase(PATCH)){
-            doPatch(request, response);
-        } else {
-            super.service(request, response);
-        }
-    }
-
-    /** Handles the PATCH request from the client. */
-    public abstract void doPatch(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException;
-}
